Move the classes used to mock system library class to conftest.py
